test: prevent editc leakage (x2)

+++ test/foundry/testUtils/PositionUtils.sol
@@ -1438,7 +1438,7 @@ contract PositionUtils is Test {
                     LeftRightSigned.unwrap(
                         LeftRightSigned
                             .wrap(int256(uint256(vm.load(address(ct), bytes32(uint256(7))))))
-                            .add(LeftRightSigned.wrap(int128(assetDelta)))
+                            .add(LeftRightSigned.wrap(int256(uint256(uint128(int128(assetDelta))))))
                     )
                 )
             )
